backend/requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
pyjwt
pydantic
pytesseract
Pillow
pymupdf
boto3
pyclamd
